// src/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]


export type Database = {
  public: {
    Tables: {
// ...
      room_agents: {
        Row: {
          agent_id: number
          created_at: string
          id: number
          room_id: number
          updated_at: string
        }
        Insert: {
          agent_id: number
          created_at?: string
          id?: number
          room_id: number
          updated_at?: string
        }
        Update: {
          agent_id?: number
          created_at?: string
          id?: number
          room_id?: number
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "room_agents_agent_id_fkey"
            columns: ["agent_id"]
            isOneToOne: false
            referencedRelation: "agents"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "room_agents_room_id_fkey"
            columns: ["room_id"]
            isOneToOne: false
            referencedRelation: "rooms"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      rooms: {
        Row: {
          active: boolean
          color: string | null
          contract_address: string | null
          created_at: string
          creator_id: number
          game_master_action_log: Json | null
          game_master_id: number | null
          id: number
          image_url: string | null
          name: string
          network: string
          pvp_action_log: Json | null
          room_config: Json | null
          round_ends_on: string | null
          round_time: number | null
          type_id: number
          updated_at: string
        }
        Insert: {
          active?: boolean
          color?: string | null
          contract_address?: string | null
          created_at?: string
          creator_id: number
          game_master_action_log?: Json | null
          game_master_id?: number | null
          id?: number
          image_url?: string | null
          name: string
          network: string
          pvp_action_log?: Json | null
          room_config?: Json | null
          round_ends_on?: string | null
          round_time?: number | null
          type_id: number
          updated_at?: string
        }
        Update: {
          active?: boolean
          color?: string | null
          contract_address?: string | null
          created_at?: string
          creator_id?: number
          game_master_action_log?: Json | null
          game_master_id?: number | null
          id?: number
          image_url?: string | null
          name?: string
          network?: string
          pvp_action_log?: Json | null
          room_config?: Json | null
          round_ends_on?: string | null
          round_time?: number | null
          type_id?: number
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "rooms_creator_id_fkey"
            columns: ["creator_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "rooms_game_master_id_fkey"
            columns: ["game_master_id"]
            isOneToOne: false
            referencedRelation: "game_masters"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "rooms_type_id_fkey"
            columns: ["type_id"]
            isOneToOne: false
            referencedRelation: "room_types"
            referencedColumns: ["id"]
          },
        ]
      }
      round_agent_messages: {
        Row: {
          agent_id: number
          created_at: string
          id: number
          message: Json
          round_id: number
          updated_at: string
        }
        Insert: {
          agent_id: number
          created_at?: string
          id?: number
          message: Json
          round_id: number
          updated_at?: string
        }
        Update: {
          agent_id?: number
          created_at?: string
          id?: number
          message?: Json
          round_id?: number
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "round_agent_messages_agent_id_fkey"
            columns: ["agent_id"]
            isOneToOne: false
            referencedRelation: "agents"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "round_agent_messages_round_id_fkey"
            columns: ["round_id"]
            isOneToOne: false
            referencedRelation: "rounds"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      rounds: {
        Row: {
          active: boolean
          created_at: string
          game_master_action_log: Json | null
          game_master_id: number | null
          id: number
          outcome: Json | null
          pvp_action_log: Json | null
          room_id: number
          round_config: Json | null
          updated_at: string
        }
        Insert: {
          active?: boolean
          created_at?: string
          game_master_action_log?: Json | null
          game_master_id?: number | null
          id?: number
          outcome?: Json | null
          pvp_action_log?: Json | null
          room_id: number
          round_config?: Json | null
          updated_at?: string
        }
        Update: {
          active?: boolean
          created_at?: string
          game_master_action_log?: Json | null
          game_master_id?: number | null
          id?: number
          outcome?: Json | null
          pvp_action_log?: Json | null
          room_id?: number
          round_config?: Json | null
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "rounds_game_master_id_fkey"
            columns: ["game_master_id"]
            isOneToOne: false
            referencedRelation: "game_masters"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "rounds_room_id_fkey"
            columns: ["room_id"]
            isOneToOne: false
            referencedRelation: "rooms"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}